refactor: rename component to components and fix import paths
