+++ packages/mesh/app/_components/category-story/main-card.tsx
