+++ packages/mesh/app/_components/daily-highlight/main-group.tsx
