+++ packages/mesh/app/_components/daily-highlight/section.tsx
