+++ packages/mesh/app/_components/daily-highlight/swiper-component.tsx
