+++ packages/mesh/app/_components/featured-card.tsx
