+++ packages/mesh/app/page.tsx
@@ -1,10 +1,10 @@
-import CategoryStorySection from './_component/category-story/section'
-import DailyHighlightSection from './_component/daily-highlight/section'
-import MostLikedCommentSection from './_component/most-liked-comments/section'
-import MostPickedStorySection from './_component/most-picked-story'
-import ReadrStorySection from './_component/readr-story'
-import TopCollectorSection from './_component/top-collector/section'
-import TopPublisherSection from './_component/top-publisher/section'
+import CategoryStorySection from './_components/category-story/section'
+import DailyHighlightSection from './_components/daily-highlight/section'
+import MostLikedCommentSection from './_components/most-liked-comments/section'
+import MostPickedStorySection from './_components/most-picked-story'
+import ReadrStorySection from './_components/readr-story'
+import TopCollectorSection from './_components/top-collector/section'
+import TopPublisherSection from './_components/top-publisher/section'
 import { getCurrentUser } from './actions/auth'
 import { fetchMemberInfo } from './actions/get-homepage'
